+++ setup.cfg
@@ -49,10 +49,10 @@ omit =
     pytorch_lightning/utilities/distributed.py
     pytorch_lightning/tuner/auto_gpu_select.py
     # TODO: temporary, until accelerator refactor is finished
-    # pytorch_lightning/accelerators/accelerator.py
-    #pytorch_lightning/plugins/training_type/*.py
-    #pytorch_lightning/plugins/precision/*.py
-    #pytorch_lightning/plugins/base_plugin.py
+    pytorch_lightning/accelerators/accelerator.py
+    pytorch_lightning/plugins/training_type/*.py
+    pytorch_lightning/plugins/precision/*.py
+    pytorch_lightning/plugins/base_plugin.py
 
 
 [flake8]
